redditScrapper/requirements.txt
praw>=7.7.0
google-generativeai>=0.8.5
python-dotenv>=1.0.1
fpdf>=1.7.2
streamlit>=1.34.0
